feat: refactor messaging components and implement ConversationItem for better code organization
